authorization implemented

+++ src/api/notes/handler.js
@@ -12,7 +12,10 @@ class NotesHandler {
   async postNoteHandler(request, h) {
     this._validator.validateNotePayload(request.payload);
     const { title = 'untitled', body, tags } = request.payload;
-    const noteId = await this._service.addNote({ title, body, tags });
+    const { id: credentialsId } = request.auth.credentials;
+    const noteId = await this._service.addNote({
+      title, body, tags, owner: credentialsId,
+    });
 
     const response = h.response({
       status: 'success',
@@ -25,8 +28,9 @@ class NotesHandler {
     return response;
   }
 
-  async getNotesHandler() {
-    const notes = await this._service.getNotes();
+  async getNotesHandler(request) {
+    const { id: credentialsId } = request.auth.credentials;
+    const notes = await this._service.getNotes(credentialsId);
     return {
       status: 'success',
       data: {
@@ -37,6 +41,8 @@ class NotesHandler {
 
   async getNoteByIdHandler(request, h) {
     const { id } = request.params;
+    const { id: credentialsId } = request.auth.credentials;
+    await this._service.verifyNoteOwner(id, credentialsId);
     const note = await this._service.getNoteById(id);
     return {
       status: 'success',
@@ -49,6 +55,8 @@ class NotesHandler {
   async putNoteByIdHandler(request, h) {
     this._validator.validateNotePayload(request.payload);
     const { id } = request.params;
+    const { id: credentialsId } = request.auth.credentials;
+    await this._service.verifyNoteOwner(id, credentialsId);
     await this._service.editNoteById(id, request.payload);
     return {
       status: 'success',
@@ -58,6 +66,8 @@ class NotesHandler {
 
   async deleteNoteByIdHandler(request, h) {
     const { id } = request.params;
+    const { id: credentialsId } = request.auth.credentials;
+    await this._service.verifyNoteOwner(id, credentialsId);
     await this._service.deleteNoteById(id);
     return {
       status: 'success',
